Use bootstrap_alert for bootstrap_messages
