+++ src/django_bootstrap5/components.py
@@ -1,3 +1,4 @@
+from django.template.defaultfilters import capfirst
 from django.utils.html import format_html
 from django.utils.translation import gettext as _
 
@@ -7,7 +8,7 @@ from .html import render_tag
 ALERT_TYPES = ["primary", "secondary", "succes", "danger", "warning", "info", "light", "dark"]
 
 
-def render_alert(content, alert_type="info", dismissible=True):
+def render_alert(content, alert_type="info", dismissible=True, extra_classes=""):
     """Render a Bootstrap alert."""
     button = ""
     if alert_type not in ALERT_TYPES:
@@ -15,7 +16,7 @@ def render_alert(content, alert_type="info", dismissible=True):
     css_classes = [f"alert alert-{alert_type}"]
     if dismissible:
         css_classes.append("alert-dismissible fade show")
-        close = _("close")
+        close = capfirst(_("close"))
         button = f'<button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="{close}"></button>'
     css_classes = merge_css_classes(*css_classes)
     return render_tag(
