+++ src/django_bootstrap5/templates/django_bootstrap5/messages.html
@@ -1,7 +1,4 @@
-{% load i18n %}{% load django_bootstrap5 %}
+{% load django_bootstrap5 %}
 {% for message in messages %}
-    <div class="{{ message|bootstrap_message_classes }} alert-dismissible fade show" role="alert">
-        <button type="button" class="close" data-dismiss="alert" aria-label="{% trans 'close' %}">&#215;</button>
-        {{ message }}
-    </div>
+    {% bootstrap_alert message alert_type=message|bootstrap_message_alert_type extra_classes=message.extra_tags %}
 {% endfor %}
