+++ tests/test_components.py
@@ -7,15 +7,13 @@ from django_bootstrap5.forms import render_button
 
 
 class AlertsTest(TestCase):
-    CLOSE = '<button aria-label="close" class="btn-close" data-bs-dismiss="alert" type="button">'
-
     def test_render_alert_without_type(self):
         self.assertEqual(
             render_alert("content"),
             (
                 '<div class="alert alert-info alert-dismissible fade show" role="alert">'
                 "content"
-                '<button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="close"></button>'
+                '<button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>'
                 "</div>"
             ),
         )
@@ -26,7 +24,7 @@ class AlertsTest(TestCase):
             (
                 '<div class="alert alert-danger alert-dismissible fade show" role="alert">'
                 "content"
-                '<button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="close"></button>'
+                '<button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>'
                 "</div>"
             ),
         )
@@ -37,7 +35,7 @@ class AlertsTest(TestCase):
             (
                 '<div class="alert alert-info alert-dismissible fade show" role="alert">'
                 'This is <a href="https://example.com" class="alert-link">a safe link</a>!'
-                '<button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="close"></button>'
+                '<button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>'
                 "</div>"
             ),
         )
@@ -48,7 +46,7 @@ class AlertsTest(TestCase):
             (
                 '<div class="alert alert-info alert-dismissible fade show" role="alert">'
                 "This is &lt;b&gt;unsafe&lt;/b&gt;!"
-                '<button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="close"></button>'
+                '<button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>'
                 "</div>"
             ),
         )
